requirements.txt
streamlit
langchain
langchain-together
langchain-huggingface
langchain-chroma
langchain-core
langchain-community
python-dotenv
PyMuPDF
arxiv
sentence-transformers